ENGR00231269: Fix renderscript on iMX53 SMD Jellybean 4.2
Make sure to disconnect the DummyConsumer to clear up the BufferQueue mQueue.
This is needed for z430 GPU on iMX53 since eglCreateWindowSurface post 2 buffers
to this default surface BufferQueue. When eglDestroySurface is called it waits
for BufferQueue mQueue to be empty. Since its a Dummy, there is no one consuming
the buffers from mQueue, hence Disconnect the consumer explicitly. Also change
the default buffer count on BufferQueue to 3 as needed by OpenGL.

Signed-off-by: Nitin Garg <nitin.garg@freescale.com>